Refactor webinars codec
