+++ .changeset/thin-files-raise.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": patch
+---
+
+Refactor webinars codec
